+++ app/scripts/components/trend-diagram.js
@@ -2,7 +2,7 @@
 import angular from "angular"
 import _ from "lodash"
 import settings from "@/settings"
-import model from "@/model"
+import model from "@/korp-api/model"
 import { expandOperators } from "@/cqp_parser/cqp"
 import { formatRelativeHits, hitCountHtml, html, loc } from "@/util"
 import * as trendUtil from "../trend_diagram/trend_util"