+++ app/scripts/controllers/word_picture_controller.js
@@ -1,7 +1,7 @@
 /** @format */
 import _ from "lodash"
 import settings from "@/settings"
-import model from "@/model"
+import model from "@/korp-api/model"
 import { isLemgram, lemgramToHtml, unregescape } from "@/util"
 
 const korpApp = angular.module("korpApp")