+++ app/scripts/parallel/stats_proxy.ts
@@ -1,9 +1,9 @@
 /** @format */
-import model from "@/model"
+import model from "@/korp-api/model"
 import settings from "@/settings"
 
 model.StatsProxy = class ParallelStatsProxy extends model.StatsProxy {
-    makeParameters(reduceVals, cqp, ignoreCase) {
+    makeParameters(reduceVals: string[], cqp: string, ignoreCase: boolean) {
         let params = super.makeParameters(reduceVals, cqp, ignoreCase)
         params.within = settings.corpusListing.getAttributeQuery("within").replace(/\|.*?:/g, ":")
         return params