+++ app/scripts/services.js
@@ -3,10 +3,11 @@ import _ from "lodash"
 import jStorage from "../lib/jstorage"
 import settings from "@/settings"
 import currentMode from "@/mode"
-import model from "@/model"
+import model from "@/korp-api/model"
 import * as authenticationProxy from "@/components/auth/auth"
 import { parseMapData } from "./map_services"
 import { updateSearchHistory } from "@/history"
+import { normalizeStatsData } from "@/korp-api/stats-proxy"
 import { httpConfAddMethod, httpConfAddMethodAngular, unregescape } from "@/util"
 import { mergeCqpExprs, parse, stringify } from "./cqp_parser/cqp"
 
@@ -209,8 +210,8 @@ korpApp.factory("backend", [
 
             return $http(conf).then(
                 function ({ data }) {
-                    model.normalizeStatsData(data)
-                    let result = parseMapData(data, cqp, cqpExprs)
+                    const normalizedData = normalizeStatsData(data)
+                    let result = parseMapData(normalizedData, cqp, cqpExprs)
                     return { corpora: attribute.corpora, cqp, within, data: result, attribute }
                 },
                 (err) => {